Move instances and services to shared
