--- src/api/index.ts
@@ -1,2 +1,0 @@
-export * from './instances';
-export * from './services';
