+++ src/modules/main/admin/dictionaries/components/dictionaries-table/confirm-dictionary-delete/ConfirmDictionaryDelete.tsx
@@ -1,8 +1,7 @@
 import React, { useState, useCallback } from 'react';
 import { useHistory } from 'react-router';
 
-import { deleteTechnology, deletePattern } from 'api';
-
+import { deleteTechnology, deletePattern } from 'shared/services';
 import { ConfirmDelete } from 'shared/components';
 
 import { DictionaryKind, Dictionary } from '../../..';
