--- src/modules/main/admin/users/models/index.ts
@@ -1,1 +1,0 @@
-export * from './users';
