--- src/modules/main/admin/users/services/index.ts
@@ -1,1 +1,0 @@
-export * from './users';
