+++ src/shared/instances/index.ts
@@ -1,0 +1,1 @@
+export * from './core';
