+++ src/shared/models/index.ts
@@ -3,5 +3,6 @@ export * from './auth';
 export * from './notifications';
 export * from './patterns';
 export * from './user';
+export * from './usersPermissions';
 export * from './templates';
 export * from './technologies';
