+++ src/shared/services/account.ts
@@ -1,6 +1,6 @@
 import { RegisterPayload, Self, ForgottenPasswordPayload } from 'shared/models';
 
-import { core } from '..';
+import { core } from 'shared/instances';
 
 const PATH = 'Account';
