+++ src/shared/services/index.ts
@@ -2,5 +2,6 @@ export * from './account';
 export * from './auth';
 export * from './notifications';
 export * from './patterns';
+export * from './usersPermissions';
 export * from './templates';
 export * from './technologies';
